requirements.txt
discord.py
python-dotenv
mistralai
autogen